Implement libraries. Loads all for now.
proxy.goL101:
	// TODO: Make libraries and endpoints many-to-many, and only use the linked ones

I think I'm of a mind to make proxy endpoints specify which libraries they want
so that we can minimize loading time. But that might also only be an issue
on deployments with lots of library code that is not used everywhere, at the cost
of UI complexity in the average case.

It also means that libraries should be able to include libraries, so that libraries
that build on other libraries can be made, which doubles the level of complexity.

Options are:
	1) All libraries, all the time
	2) Only linked libraries
	3) Config option to toggle UI/behavior
	4) Something else

A something else could be a code level:

	AP.includeLibrary("library name");

Which is good and bad. It's good because it's clear and explicit; it's bad because
function calls should report errors back when bad stuff happens, and it's not something
a developer would want to be checking at that state. Maybe I can make it return a
runtime error to do that.

Might leave it alone for now... with a ticket to evaluate more in the future filed
here: https://www.pivotaltracker.com/story/show/82001452

[Finishes #81308456]