Prefer to ignore generated files.
Golang wisdom from above is to check in generated files, as
future authors might not have the commands to execute, etc.
However, in our case I am checking in the scripts (so far),
and we're in control of the limited development environment.
And it makes my life easier as I can just ignore _gen files
in my text editor.

So there.