Add SQL Server Remote Endpoints and db package
`proxy/Server` now adds a `gateway/db/pools/Pools` connection cache with
a pool prepared for SQL Server connections.  This Pools type can easily
be extended with new database implementations as needed.  It offers a
"friendly" functional API as defined in
http://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis .

When a Request comes in, `proxy.prepareRequest` generates a SQL Server or HTTP
Request implementation depending on the `Type` in the Request's `Data` field.
If a SQL Server request is prepared, it will call Connect on the
`proxy/Server.dbPools`.  This will use a cached connection if one is
available.

Transactions can also be enabled for the Request using the `Transactions`
field of `SQLServerRequest` but are disabled by default.