[Fix #98753086, #96902506] Check before clobbering vm Interrupt channel.
Each time Run is called on a vm, it replaces the vm's Interrupt channel
with a new one.  This means the timeout is effectively reset each time
Run is called.  We now check before creating the new Interrupt channel.

This was also causing a memory leak by retaining the goroutine
containing the Interrupt send indefinitely.  `pprof` was used to show
that the number of goroutines is no longer increasing without bound.